Initial commit of real, actual code.
Kinda like pinochio, the real boy that didn't start his life as a real boy.  This code actually works and has unit tests and an "integration" thingie in example/integrated.js.  The "integration" thingie is, once again, more complex than I think I'd like, but it's exercising things going up and down...  Can be iterated on in the future.